+++ src/main/relations/Relation.java
@@ -7,9 +7,9 @@ public class Relation {
     public Element el1;
     public Element el2;
     public int relId;
-    public float value;
+    public double value;
 
-    public Relation(Element el1, Element el2, int relId, float value) {
+    public Relation(Element el1, Element el2, int relId, double value) {
         this.el1 = el1;
         this.el2 = el2;
         this.relId = relId;
